# docs/swagger.yaml
basePath: /api
definitions:
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@anheyu.com
    name: 安知鱼
    url: https://github.com/anzhiyu-c/anheyu-app
  description: Anheyu App 应用接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Anheyu App API
  version: "1.0"
paths:
...
  /t/{signedToken}:
    get:
      description: 通过签名令牌访问缩略图文件内容
      parameters:
      - description: 签名令牌
        in: path
        name: signedToken
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 缩略图内容
          schema:
            type: file
        "403":
          description: 签名无效或已过期
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 资源未找到
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 提供内容失败
          schema:
            $ref: '#/definitions/response.Response'
      summary: 访问缩略图内容
      tags:
      - 缩略管理
...
  /thumbnail/regenerate:
    post:
      consumes:
      - application/json
      description: 手动触发单个文件的缩略图重新生成
      parameters:
      - description: 文件公共ID
        in: body
        name: body
        required: true
        schema:
          properties:
            id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "202":
          description: 任务已启动
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: 访问被拒绝
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 文件未找到
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 重置失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 重新生成缩略图
      tags:
      - 缩略管理
  /thumbnail/regenerate-directory:
    post:
      consumes:
      - application/json
      description: 为指定目录下的所有文件重新生成缩略图
      parameters:
      - description: 目录公共ID
        in: body
        name: body
        required: true
        schema:
          properties:
            directoryId:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "202":
          description: 批量任务已启动
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数无效或目标不是文件夹
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 目录未找到
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 获取文件列表失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 批量重新生成缩略图
      tags:
      - 缩略管理
  /thumbnail/sign/{publicID}:
    get:
      description: 获取文件缩略图的访问签名，如果缩略图未就绪会触发生成
      parameters:
      - description: 文件公共ID
        in: path
        name: publicID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 签名获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  properties:
                    expires:
                      type: string
                    obfuscated:
                      type: boolean
                    sign:
                      type: string
                  type: object
              type: object
        "202":
          description: 缩略图生成中
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  properties:
                    status:
                      type: string
                  type: object
              type: object
        "400":
          description: 无效的文件ID
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: 访问被拒绝
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 文件未找到
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取缩略图签名
      tags:
      - 缩略管理
...
securityDefinitions:
  BearerAuth:
    description: '在请求头中添加 Bearer Token，格式为: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"